+++ R-package/src/ball_init.c
@@ -9,8 +9,6 @@ Check these declarations against the C/Fortran source code.
 extern void bcor_test(double *, double *, double *, int *, int *, int *, int *, int *, int *, int *, int *);
 extern void bcov_test(double *, double *, double *, double *, int *, int *, int *, int *);
 extern void kbcov_test(double *, double *, double *, int *, int *, int *, int *, int *);
-extern void bd_gwas_screening(double *, double *, double *, int *, int *, double *, int *, int *, int *, int *, int *, int *, int *, int *);
-extern void bd_gwas_refining_single(double *, double *, double *, int *, int *, double *, int *, int *, int *, int *, int *, int *, int *, int *);
 extern void bd_test(double *, double *, double *, int *, int *, int *, int *, int *, int *);
 extern void SRCT_new(double *, int *, int *, double *, int *, double *);
 extern void bdd_matrix_bias(double *, double *, int *, int *);
@@ -20,8 +18,6 @@ static const R_CMethodDef CEntries[] = {
   {"bcor_test", (DL_FUNC) &bcor_test, 11},
   {"bcov_test", (DL_FUNC) &bcov_test, 8},
   {"kbcov_test", (DL_FUNC) &kbcov_test, 8},
-  {"bd_gwas_refining_single", (DL_FUNC) &bd_gwas_refining_single, 14},
-  {"bd_gwas_screening",       (DL_FUNC) &bd_gwas_screening,       14},
   {"bd_test",   (DL_FUNC) &bd_test,   9},
   {"SRCT_new",  (DL_FUNC) &SRCT_new,  6},
   {"bdd_matrix_bias",           (DL_FUNC) &bdd_matrix_bias,           4},